ironic-oneviewd/requirements.txt
pbr>=1.6 # Apache-2.0
oslo.config>=3.14.0 # Apache-2.0
oslo.i18n>=2.1.0 # Apache-2.0
oslo.log>=3.11.0 # Apache-2.0
oslo.utils>=3.16.0 # Apache-2.0
python-ironicclient>=1.6 # Apache-2.0
hpOneView>=3.2.1
requests!=2.9.0,>=2.8.1
future
retrying>=1.2.3
futures
